Working version for single_table validation
