+++ tests/unit/metadata/test_single_table.py
@@ -2,12 +2,14 @@
 
 import json
 import re
+from datetime import datetime
 from unittest.mock import Mock, call, patch
 
 import numpy as np
 import pandas as pd
 import pytest
 
+from sdv.errors import InvalidDataError
 from sdv.metadata.errors import InvalidMetadataError
 from sdv.metadata.single_table import SingleTableMetadata
 
@@ -320,14 +322,14 @@ class TestSingleTableMetadata:
             'supported SDV sdtypes.'
         )
         with pytest.raises(InvalidMetadataError, match=error_msg):
-            instance._validate_column('column', 'fake_type')
+            instance._validate_column_args('column', 'fake_type')
 
         error_msg = re.escape(
             'Invalid sdtype: None is not a string. Please use one of the '
             'supported SDV sdtypes.'
         )
         with pytest.raises(InvalidMetadataError, match=error_msg):
-            instance._validate_column('column', None)
+            instance._validate_column_args('column', None)
 
     @patch('sdv.metadata.single_table.SingleTableMetadata._validate_unexpected_kwargs')
     @patch('sdv.metadata.single_table.SingleTableMetadata._validate_numerical')
@@ -355,7 +357,7 @@ class TestSingleTableMetadata:
         instance = SingleTableMetadata()
 
         # Run
-        instance._validate_column('age', 'numerical', computer_representation='Int8')
+        instance._validate_column_args('age', 'numerical', computer_representation='Int8')
 
         # Assert
         mock__validate_kwargs.assert_called_once_with(
@@ -388,7 +390,7 @@ class TestSingleTableMetadata:
         instance = SingleTableMetadata()
 
         # Run
-        instance._validate_column('name', 'categorical', order=['a', 'b', 'c'])
+        instance._validate_column_args('name', 'categorical', order=['a', 'b', 'c'])
 
         # Assert
         mock__validate_kwargs.assert_called_once_with(
@@ -418,7 +420,7 @@ class TestSingleTableMetadata:
         instance = SingleTableMetadata()
 
         # Run
-        instance._validate_column('snythetic', 'boolean')
+        instance._validate_column_args('snythetic', 'boolean')
 
         # Assert
         mock__validate_kwargs.assert_called_once_with('snythetic', 'boolean')
@@ -449,7 +451,7 @@ class TestSingleTableMetadata:
         instance = SingleTableMetadata()
 
         # Run
-        instance._validate_column('start', 'datetime')
+        instance._validate_column_args('start', 'datetime')
 
         # Assert
         mock__validate_kwargs.assert_called_once_with('start', 'datetime')
@@ -481,7 +483,7 @@ class TestSingleTableMetadata:
         instance = SingleTableMetadata()
 
         # Run
-        instance._validate_column('phrase', 'id', regex_format='[A-z0-9]', pii=True)
+        instance._validate_column_args('phrase', 'id', regex_format='[A-z0-9]', pii=True)
 
         # Assert
         mock__validate_kwargs.assert_called_once_with(
@@ -647,10 +649,10 @@ class TestSingleTableMetadata:
         # Assert
         assert instance.columns['number'] == {'sdtype': 'phone_number', 'pii': True}
 
-    @patch('sdv.metadata.single_table.SingleTableMetadata._validate_column')
+    @patch('sdv.metadata.single_table.SingleTableMetadata._validate_column_args')
     @patch('sdv.metadata.single_table.SingleTableMetadata._validate_column_exists')
-    def test_upate_column_sdtype_in_kwargs(self,
-                                           mock__validate_column_exists, mock__validate_column):
+    def test_update_column_sdtype_in_kwargs(self,
+                                            mock__validate_column_exists, mock__validate_column):
         """Test the ``update_column`` method.
 
         Test that when calling ``update_column`` with an ``sdtype`` this is being updated as well
@@ -683,9 +685,9 @@ class TestSingleTableMetadata:
         mock__validate_column.assert_called_once_with(
             'age', 'categorical', order_by='numerical_value')
 
-    @patch('sdv.metadata.single_table.SingleTableMetadata._validate_column')
+    @patch('sdv.metadata.single_table.SingleTableMetadata._validate_column_args')
     @patch('sdv.metadata.single_table.SingleTableMetadata._validate_column_exists')
-    def test_upate_column_no_sdtype(self, mock__validate_column_exists, mock__validate_column):
+    def test_update_column_no_sdtype(self, mock__validate_column_exists, mock__validate_column):
         """Test the ``update_column`` method.
 
         Test that when calling ``update_column`` without an ``sdtype`` is updating the other
@@ -1465,7 +1467,7 @@ class TestSingleTableMetadata:
         instance._validate_alternate_keys = Mock()
         instance._validate_sequence_index = Mock()
         instance._validate_sequence_index_not_in_sequence_key = Mock()
-        instance._validate_column = Mock(side_effect=InvalidMetadataError('column_error'))
+        instance._validate_column_args = Mock(side_effect=InvalidMetadataError('column_error'))
 
         err_msg = re.escape(
             'The following errors were found in the metadata:'
@@ -1480,12 +1482,332 @@ class TestSingleTableMetadata:
         instance._validate_key.assert_has_calls(
             [call(instance.primary_key, 'primary'), call(instance.sequence_key, 'sequence')]
         )
-        instance._validate_column.assert_has_calls(
+        instance._validate_column_args.assert_has_calls(
             [call('col1', sdtype='numerical'), call('col2', sdtype='numerical')]
         )
         instance._validate_alternate_keys.assert_called_once_with(instance.alternate_keys)
         instance._validate_sequence_index.assert_called_once_with(instance.sequence_index)
         instance._validate_sequence_index_not_in_sequence_key.assert_called_once()
+
+    def test_validate_with_data_wrong_type(self):
+        """Test error is raised if data is not ``pd.DataFrame``."""
+        # Setup
+        data = np.ndarray([])
+        metadata = SingleTableMetadata()
+
+        # Run and Assert
+        err_msg = "Data must be a DataFrame, not a <class 'numpy.ndarray'>."
+        with pytest.raises(ValueError, match=err_msg):
+            metadata.validate_with_data(data)
+
+    def test_validate_with_data_data_columns_in_empty_metadata(self):
+        """Test error is raised if data is passed and metadata is empty."""
+        # Setup
+        data = pd.DataFrame({
+            'col1': [1, 2, 3],
+            'col2': [4, 5, 6],
+        })
+        metadata = SingleTableMetadata()
+
+        # Run and Assert
+        err_msg = re.escape(
+            'The provided data does not match the metadata:'
+            "\nThe columns ['col1', 'col2'] are not present in the metadata."
+        )
+        with pytest.raises(InvalidDataError, match=err_msg):
+            metadata.validate_with_data(data)
+
+    def test_validate_with_data_data_columns_in_metadata(self):
+        """Test error is raised if data columns don't match metadata columns."""
+        # Setup
+        data = pd.DataFrame({
+            'col1': [1, 2, 3],
+            'col2': [4, 5, 6],
+            'col3': [7, 8, 9],
+        })
+        metadata = SingleTableMetadata()
+        metadata.add_column('col1', sdtype='numerical')
+        metadata.add_column('col4', sdtype='numerical')
+        metadata.add_column('col5', sdtype='numerical')
+
+        # Run and Assert
+        err_msg = re.escape(
+            'The provided data does not match the metadata:'
+            "\nThe columns ['col2', 'col3'] are not present in the metadata."
+            '\n'
+            "\nThe metadata columns ['col4', 'col5'] are not present in the data."
+        )
+        with pytest.raises(InvalidDataError, match=err_msg):
+            metadata.validate_with_data(data)
+
+    def test_validate_with_data_keys_with_missing_values(self):
+        """Test error is raised if keys contain missing values.
+
+        Setup:
+            A ``SingleTableMetadata`` instance with one primary key and multiple sequence
+            and alternate keys. All the columns contain missing values except for one
+            squence key and one alternate key, so we can ensure those don't show up
+            in the error message.
+        """
+        data = pd.DataFrame({
+            'pk_col': [0, 1, np.nan],
+            'sk_col1': [0, 1, None],
+            'sk_col2': [0, 1, np.nan],
+            'sk_col3': [0, 1, 2],
+            'ak_col1': [0, 1, None],
+            'ak_col2': [0, 1, np.nan],
+            'ak_col3': [0, 1, 2]
+        })
+        metadata = SingleTableMetadata()
+        metadata.add_column('pk_col', sdtype='id')
+        metadata.add_column('sk_col1', sdtype='id')
+        metadata.add_column('sk_col2', sdtype='id')
+        metadata.add_column('sk_col3', sdtype='id')
+        metadata.add_column('ak_col1', sdtype='id')
+        metadata.add_column('ak_col2', sdtype='id')
+        metadata.add_column('ak_col3', sdtype='id')
+        metadata.set_primary_key('pk_col')
+        metadata.set_sequence_key(('sk_col1', 'sk_col2', 'sk_col3'))
+        metadata.add_alternate_keys(['ak_col1', 'ak_col2', 'ak_col3'])
+
+        # Run and Assert
+        err_msg = re.escape(
+            'The provided data does not match the metadata:'
+            "\nKey column 'ak_col1' contains missing values."
+            '\n'
+            "\nKey column 'ak_col2' contains missing values."
+            '\n'
+            "\nKey column 'pk_col' contains missing values."
+            '\n'
+            "\nKey column 'sk_col1' contains missing values."
+            '\n'
+            "\nKey column 'sk_col2' contains missing values."
+        )
+        with pytest.raises(InvalidDataError, match=err_msg):
+            metadata.validate_with_data(data)
+
+    def test_validate_with_data_keys_with_missing_with_single_sequence_key(self):
+        """Test error is raised if keys contain missing values.
+
+        Test the case with a single sequence key.
+        """
+        data = pd.DataFrame({
+            'pk_col': [1],
+            'sk_col': [None]
+        })
+        metadata = SingleTableMetadata()
+        metadata.add_column('pk_col', sdtype='id')
+        metadata.add_column('sk_col', sdtype='id')
+        metadata.set_primary_key('pk_col')
+        metadata.set_sequence_key('sk_col')
+
+        # Run and Assert
+        err_msg = re.escape(
+            'The provided data does not match the metadata:'
+            "\nKey column 'sk_col' contains missing values."
+        )
+        with pytest.raises(InvalidDataError, match=err_msg):
+            metadata.validate_with_data(data)
+
+    def test_validate_with_data_keys_not_unique(self):
+        """Test error is raised if primary or alternate keys are not unique."""
+        data = pd.DataFrame({
+            'pk_col': [0, 1, 1, 0, 2],
+            'ak_col1': [0, 1, 0, 3, 3],
+            'ak_col2': [2, 2, 2, 2, 2],
+            'ak_col3': [0, 1, 2, 3, 4]
+        })
+        metadata = SingleTableMetadata()
+        metadata.add_column('pk_col', sdtype='id')
+        metadata.add_column('ak_col1', sdtype='id')
+        metadata.add_column('ak_col2', sdtype='id')
+        metadata.add_column('ak_col3', sdtype='id')
+        metadata.set_primary_key('pk_col')
+        metadata.add_alternate_keys(['ak_col1', 'ak_col2', 'ak_col3'])
+
+        # Run and Assert
+        err_msg = re.escape(
+            'The provided data does not match the metadata:'
+            "\nKey column 'ak_col1' contains repeating values: [0, 3]"
+            '\n'
+            "\nKey column 'ak_col2' contains repeating values: [2]"
+            '\n'
+            "\nKey column 'pk_col' contains repeating values: [0, 1]"
+        )
+        with pytest.raises(InvalidDataError, match=err_msg):
+            metadata.validate_with_data(data)
+
+    def test_validate_with_data_empty(self):
+        """Test method doesn't raise when data is empty.
+
+        Setup:
+            ``SingleTableMetadata`` with one column for each sdtype and for each key.
+        """
+        data = pd.DataFrame({
+            'pk_col': [],
+            'sk_col': [],
+            'ak_col': [],
+            'bool_col': [],
+            'num_col': [],
+            'date_col': [],
+        })
+        metadata = SingleTableMetadata()
+        metadata.add_column('bool_col', sdtype='boolean')
+        metadata.add_column('num_col', sdtype='numerical')
+        metadata.add_column('date_col', sdtype='datetime')
+        metadata.add_column('pk_col', sdtype='id')
+        metadata.add_column('sk_col', sdtype='id')
+        metadata.add_column('ak_col', sdtype='id')
+        metadata.set_primary_key('pk_col')
+        metadata.set_sequence_key('sk_col')
+        metadata.add_alternate_keys(['ak_col'])
+
+        # Run
+        metadata.validate_with_data(data)
+
+    def test_validate_with_data_no_keys(self):
+        """Test method passes even if no keys are passed."""
+        data = pd.DataFrame({
+            'bool_col': [1, 2, 3],
+            'num_col': [1, 2, 3],
+            'date_col': [1, 2, 3],
+        })
+        metadata = SingleTableMetadata()
+        metadata.add_column('bool_col', sdtype='numerical')
+        metadata.add_column('num_col', sdtype='numerical')
+        metadata.add_column('date_col', sdtype='numerical')
+
+        # Run
+        metadata.validate_with_data(data)
+
+    def test_validate_with_data_empty_dataframe(self):
+        """Test method doesn't raise when data is an empty dataframe."""
+        data = pd.DataFrame()
+        metadata = SingleTableMetadata()
+
+        # Run
+        metadata.validate_with_data(data)
+
+    def test_validate_with_data_sdtypes(self):
+        """Test error is raised if column values don't satisfy their sdtype.
+
+        Setup:
+            A ``SingleTableMetadata`` instance with two columns of each sdtype: numerical,
+            boolean and datetime. The first column of each will have 4 invalid values,
+            while the second column will have at most 3.
+        """
+        # Setup
+        data = pd.DataFrame({
+            'date1': ['10', True, 'b', 'bla', None],
+            'date2': ['2021-10-10', '05-10-2021', pd.Timestamp(1), datetime(1, 1, 1), '2020-1-33'],
+            'bool1': ['a', 0, '10', True, 'b'],
+            'bool2': ['True', False, np.nan, float('nan'), None],
+            'num1': ['a', 0, '10', True, False],
+            'num2': [-1.2, datetime(1, 1, 1), np.nan, float('nan'), None],
+        })
+        metadata = SingleTableMetadata()
+        metadata.add_column('date1', sdtype='datetime')
+        metadata.add_column('date2', sdtype='datetime')
+        metadata.add_column('bool1', sdtype='boolean')
+        metadata.add_column('bool2', sdtype='boolean')
+        metadata.add_column('num1', sdtype='numerical')
+        metadata.add_column('num2', sdtype='numerical')
+
+        # Run and Assert
+        err_msg = re.escape(
+            'The provided data does not match the metadata:'
+            "\nInvalid values found for datetime column 'date1': ['10', True, 'b', '+ 1 more']."
+            '\n'
+            "\nInvalid values found for datetime column 'date2': ['2020-1-33']."
+            '\n'
+            "\nInvalid values found for boolean column 'bool1': [0, '10', 'a', '+ 1 more']."
+            '\n'
+            "\nInvalid values found for boolean column 'bool2': ['True']."
+            '\n'
+            "\nInvalid values found for numerical column 'num1': ['10', False, True, '+ 1 more']."
+            '\n'
+            "\nInvalid values found for numerical column 'num2': "
+            '[datetime.datetime(1, 1, 1, 0, 0)].'
+        )
+        with pytest.raises(InvalidDataError, match=err_msg):
+            metadata.validate_with_data(data)
+
+    def test_validate_with_data_datetime_sdtype(self):
+        """Test validation for columns with datetime format.
+
+        If the datetime format is provided, then the values must match. Otherwise an error should
+        be raised.
+        """
+        # Setup
+        data = pd.DataFrame({
+            'date_str': [
+                '20220902110443000000',
+                '20220916230356000000',
+                '20220826173917000000',
+                '20220826212135000000',
+                '20220929111311000000'
+            ],
+            'date_int': [
+                20220902110443000000,
+                20220916230356000000,
+                20220826173917000000,
+                20220826212135000000,
+                20220929111311000000
+            ],
+            'bad_date': [
+                2022090,
+                20220916230356000000,
+                2022,
+                20220826212135000000,
+                20220929111311000000
+            ]
+        })
+        metadata = SingleTableMetadata()
+        metadata.add_column('date_str', sdtype='datetime', datetime_format='%Y%m%d%H%M%S%f')
+        metadata.add_column('date_int', sdtype='datetime', datetime_format='%Y%m%d%H%M%S%f')
+        metadata.add_column('bad_date', sdtype='datetime', datetime_format='%Y%m%d%H%M%S%f')
+
+        # Run and Assert
+        err_msg = re.escape(
+            'The provided data does not match the metadata:'
+            "\nInvalid values found for datetime column 'bad_date': [2022, 2022090]."
+        )
+        with pytest.raises(InvalidDataError, match=err_msg):
+            metadata.validate_with_data(data)
+
+    def test_validate_with_data(self):
+        """Test the method doesn't crash when the passed data is valid.
+
+        Setup:
+            ``SingleTableMetadata`` describing at least one valid column of each key and sdtype.
+        """
+        # Setup
+        data = pd.DataFrame({
+            'pk_col': [0, 1, 2],
+            'sk_col1': [0, 1, 2],
+            'sk_col2': [0, 1, 2],
+            'ak_col1': [0, 1, 2],
+            'ak_col2': [0, 1, 2],
+            'numerical_col': [np.nan, -1, 1.54],
+            'date_col': [np.nan, '2021-02-10', '2021-05-10'],
+            'bool_col': [np.nan, True, False],
+
+        })
+        metadata = SingleTableMetadata()
+        metadata.add_column('pk_col', sdtype='id')
+        metadata.add_column('sk_col1', sdtype='id')
+        metadata.add_column('sk_col2', sdtype='id')
+        metadata.add_column('ak_col1', sdtype='id')
+        metadata.add_column('ak_col2', sdtype='id')
+        metadata.add_column('numerical_col', sdtype='numerical')
+        metadata.add_column('date_col', sdtype='datetime')
+        metadata.add_column('bool_col', sdtype='boolean')
+        metadata.set_primary_key('pk_col')
+        metadata.set_sequence_key(('sk_col1', 'sk_col2'))
+        metadata.add_alternate_keys(['ak_col1', 'ak_col2'])
+
+        # Run
+        metadata.validate_with_data(data)
 
     def test_to_dict(self):
         """Test the ``to_dict`` method from ``SingleTableMetadata``.
